\name{BCI-package}
\alias{BCI-package}
\alias{BCI}
\docType{package}
\title{
\packageTitle{BCI}
}
\description{
\packageDescription{BCI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BCI}
\packageIndices{BCI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BCI}

Maintainer: \packageMaintainer{BCI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[boot:boot-package]{boot}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
