% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bci_100.R
\docType{data}
\name{bci_100}
\alias{bci_100}
\title{BCI 50 ha plot census data with 100m square quadrats}
\format{A tbl data_frame}
\source{
https://repository.si.edu/bitstream/handle/10088/20925/bci.full.Rdata31Aug2012.zip
}
\usage{
bci_100
}
\description{
A table with one record per tree species per grid square per census,
detailing the count of the species in that grid square for that census. In
this case we use 100m square quadrats.
}
\examples{
  library(tibble)
  library(magrittr)
  library(dplyr)

  # How many trees in each census?
  bci_100 \%>\% group_by(census) \%>\% summarise(total=sum(count))

  # Look at the first quadrat of the first census
  bci_100 \%>\% filter(x==0, y==0, census==1)
}
\keyword{datasets}
