% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bci_taxonomy.R
\docType{data}
\name{bci_taxonomy}
\alias{bci_taxonomy}
\title{BCI 50 ha plot taxonomy}
\format{A tbl data_frame}
\source{
https://repository.si.edu/bitstream/handle/10088/20925/bci.full.Rdata31Aug2012.zip
}
\usage{
bci_taxonomy
}
\description{
A table with one record per tree species detailing all of the taxonomic
information about those species.
}
\examples{
  library(tibble)

  # Show the taxonomy
  bci_taxonomy
}
\keyword{datasets}
