% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zmatrix.R
\name{gen_zmatrix}
\alias{gen_zmatrix}
\title{Generate similarity matrix for a given taxonomy}
\usage{
gen_zmatrix(tax = BCI::bci_taxonomy, species = 1, genus = 0.75,
  family = 0.5, order = 0.25, other = 0)
}
\arguments{
\item{tax}{The taxonomy to use}

\item{species}{The similarity value if two rows are the same species}

\item{genus}{The similarity value if two rows are the same genus but not species}

\item{family}{The similarity value if two rows are the same family but not genus}

\item{order}{The similarity value if two rows are the same order but not family}

\item{other}{The similarity value if two rows are not the same order}
}
\value{
A matrix of similarities between species based on their taxonomy.
}
\description{

}
\examples{

# Generate the similarity matrix for the BCI dataset
gen_zmatrix()
}
