% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Cast the BCI dataset so it is usable by the rdiversity package}
\usage{
tidy_data(dataset)
}
\arguments{
\item{dataset}{The dataset to cast}
}
\value{
A tibble / dataframe with rows as species and columns as subcommunities
}
\description{

}
\examples{
library(magrittr)

# Generate a data frame containing linear 100m-square subcommunities running along the y-axis
bci_100 \%>\% dplyr::filter(census==1, x==0) \%>\% tidy_data
}
